refactor(corpus)+docs: rename layout.ts->paths.ts (Next reserved name); seam docs + OPFS smoke (per plan-corpus-s1 steps 9-10)
- Rename app/lib/corpus/layout.ts -> paths.ts: Next.js App Router treats any
  app/**/layout.ts as a route layout, failing the build with a missing-default
  error. paths.test.ts import updated.
- opfsAdapter: narrow local OpfsWritable.write to Uint8Array (TS lib
  BufferSource/ArrayBufferLike strictness).
- docs/thoughts/corpus-fs-seam.md: seam notes + S1 blob-mode/no-migration facts
  + the two build gotchas, for S2-S4.
- docs/spikes/corpus-opfs-smoke.md: out-of-CI Playwright OPFS smoke (pre-merge
  gate before the flag is defaulted on; Safari still pending).
- CLAUDE.md: document lib/corpus/ module, the dev-only flag, jsdom-no-OPFS caveat.
- plan: Actual context cost filled; Task status -> complete.

Verified: npm run lint clean, npm run build passes, 324 vitest tests pass.

Confidence: high
Notes: layout.ts reserved-name collision is a real gotcha recorded in the seam
doc + CLAUDE.md.
